GCG-28: As a visitor, I want to see reviews on a profile so that I can assess credibility.
refactor: extract date formatter into composables
