+++ resources/js/composables/useDateFormatter.ts
@@ -1,0 +1,13 @@
+export function useDateFormatter() {
+    const formatDate = (dateString: string) => {
+        return new Date(dateString).toLocaleDateString('en-US', {
+            year: 'numeric',
+            month: 'short',
+            day: 'numeric',
+        });
+    };
+
+    return {
+        formatDate
+    }
+}
